requirements.txt
opencv-python
numpy-stl
